requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
flatbuffers==24.3.7
fsspec==2024.2.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
imgbeddings==0.1.0
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
matplotlib-inline==0.1.6
mpmath==1.3.0
mysql-connector-python==8.2.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
